Sample-DataSet-CommercialLending/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
pandas
fastapi
uvicorn
python-dotenv